chore: uses from Spring configured `ObjectMapper` also for RPC
- Creates the `ObjectMapper` with `Jackson2ObjectMapperBuilder` now and thus with the general Spring configuration.
- Moves the `StatusDeserializer` to the package where this is relevant.
- code cleanup
